src/requirements.txt
tensorflow
django
djangorestframework
numpy
pandas
tensorflow
xlsxwriter